clean up and refactor
