+++ tests/execution_engine/conftest.py
@@ -210,3 +210,19 @@ def test_sparkdf(spark_session):
         "timestamp", F.col("timestamp").cast(IntegerType()).cast(StringType())
     )
     return spark_df
+
+
+@pytest.fixture
+def test_folder_connection_path_tsv(tmp_path_factory):
+    df1 = pd.DataFrame({"col_1": [1, 2, 3, 4, 5], "col_2": ["a", "b", "c", "d", "e"]})
+    path = str(tmp_path_factory.mktemp("test_folder_connection_path_tsv"))
+    df1.to_csv(path_or_buf=os.path.join(path, "test.tsv"), sep="\t", index=False)
+    return str(path)
+
+
+@pytest.fixture
+def test_folder_connection_path_parquet(tmp_path_factory):
+    df1 = pd.DataFrame({"col_1": [1, 2, 3, 4, 5], "col_2": ["a", "b", "c", "d", "e"]})
+    path = str(tmp_path_factory.mktemp("test_folder_connection_path_parquet"))
+    df1.to_parquet(path=os.path.join(path, "test.parquet"))
+    return str(path)
