+++ tests/execution_engine/split_and_sample/test_pandas_execution_engine_sampling.py
@@ -1,3 +1,6 @@
+import datetime
+import random
+
 import pandas as pd
 import pytest
 
@@ -52,3 +55,70 @@ def test_limit_sampler_get_batch_data(
     )
 
     assert len(sampled_df) == num_sampled_rows
+
+
+def test_sample_using_random(test_df):
+    random.seed(1)
+    sampled_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(batch_data=test_df, sampling_method="_sample_using_random")
+    )
+    assert sampled_df.dataframe.shape == (13, 10)
+
+
+def test_sample_using_mod(test_df):
+    sampled_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            sampling_method="_sample_using_mod",
+            sampling_kwargs={
+                "column_name": "id",
+                "mod": 5,
+                "value": 4,
+            },
+        )
+    )
+    assert sampled_df.dataframe.shape == (24, 10)
+
+
+def test_sample_using_a_list(test_df):
+    sampled_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            sampling_method="_sample_using_a_list",
+            sampling_kwargs={
+                "column_name": "id",
+                "value_list": [3, 5, 7, 11],
+            },
+        )
+    )
+    assert sampled_df.dataframe.shape == (4, 10)
+
+
+def test_sample_using_md5(test_df):
+    with pytest.raises(ge_exceptions.ExecutionEngineError):
+        # noinspection PyUnusedLocal
+        sampled_df = PandasExecutionEngine().get_batch_data(
+            RuntimeDataBatchSpec(
+                batch_data=test_df,
+                sampling_method="_sample_using_hash",
+                sampling_kwargs={
+                    "column_name": "date",
+                    "hash_function_name": "I_am_not_valid",
+                },
+            )
+        )
+
+    sampled_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            sampling_method="_sample_using_hash",
+            sampling_kwargs={"column_name": "date", "hash_function_name": "md5"},
+        )
+    )
+    assert sampled_df.dataframe.shape == (10, 10)
+    assert sampled_df.dataframe.date.isin(
+        [
+            datetime.date(2020, 1, 15),
+            datetime.date(2020, 1, 29),
+        ]
+    ).all()
