+++ tests/execution_engine/split_and_sample/test_sparkdf_execution_engine_sampling.py
@@ -1,3 +1,5 @@
+import datetime
+
 import pytest
 
 import great_expectations.exceptions as ge_exceptions
@@ -53,3 +55,92 @@ def test_limit_sampler_get_batch_data(
     ).dataframe
 
     assert sampled_df.count() == num_sampled_rows
+
+
+def test_get_batch_empty_sampler(test_sparkdf, basic_spark_df_execution_engine):
+    sampled_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(batch_data=test_sparkdf, sampling_method=None)
+    ).dataframe
+    assert sampled_df.count() == 120
+    assert len(sampled_df.columns) == 10
+
+
+def test_sample_using_random(test_sparkdf, basic_spark_df_execution_engine):
+    sampled_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf, sampling_method="_sample_using_random"
+        )
+    ).dataframe
+    # The test dataframe contains 10 columns and 120 rows.
+    assert len(sampled_df.columns) == 10
+    assert 0 <= sampled_df.count() <= 120
+    # The sampling probability "p" used in "SparkDFExecutionEngine._sample_using_random()" is 0.1 (the equivalent of an
+    # unfair coin with the 10% chance of coming up as "heads").  Hence, we should never get as much as 20% of the rows.
+    assert sampled_df.count() < 25
+
+
+def test_sample_using_mod(test_sparkdf, basic_spark_df_execution_engine):
+    sampled_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            sampling_method="_sample_using_mod",
+            sampling_kwargs={
+                "column_name": "id",
+                "mod": 5,
+                "value": 4,
+            },
+        )
+    ).dataframe
+    assert sampled_df.count() == 24
+    assert len(sampled_df.columns) == 10
+
+
+def test_sample_using_a_list(test_sparkdf, basic_spark_df_execution_engine):
+    sampled_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            sampling_method="_sample_using_a_list",
+            sampling_kwargs={
+                "column_name": "id",
+                "value_list": [3, 5, 7, 11],
+            },
+        )
+    ).dataframe
+    assert sampled_df.count() == 4
+    assert len(sampled_df.columns) == 10
+
+
+def test_sample_using_md5_wrong_hash_function_name(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    with pytest.raises(ge_exceptions.ExecutionEngineError):
+        # noinspection PyUnusedLocal
+        sampled_df = basic_spark_df_execution_engine.get_batch_data(
+            RuntimeDataBatchSpec(
+                batch_data=test_sparkdf,
+                sampling_method="_sample_using_hash",
+                sampling_kwargs={
+                    "column_name": "date",
+                    "hash_function_name": "I_wont_work",
+                },
+            )
+        ).dataframe
+
+
+def test_sample_using_md5(test_sparkdf, basic_spark_df_execution_engine):
+    sampled_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            sampling_method="_sample_using_hash",
+            sampling_kwargs={
+                "column_name": "date",
+                "hash_function_name": "md5",
+            },
+        )
+    ).dataframe
+    assert sampled_df.count() == 10
+    assert len(sampled_df.columns) == 10
+
+    collected = sampled_df.collect()
+    for val in collected:
+        assert val.date in [datetime.date(2020, 1, 15), datetime.date(2020, 1, 29)]
