+++ tests/execution_engine/split_and_sample/test_sparkdf_execution_engine_splitting.py
@@ -276,6 +276,54 @@ def test_get_splitter_method(underscore_prefix: str, splitter_method_name: str):
     ) == getattr(data_splitter, splitter_method_name)
 
 
+def test_get_batch_empty_splitter(
+    test_folder_connection_path_csv, basic_spark_df_execution_engine
+):
+    # reader_method not configured because spark will configure own reader by default
+    # reader_options are needed to specify the fact that the first line of test file is the header
+    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
+        PathBatchSpec(
+            path=os.path.join(test_folder_connection_path_csv, "test.csv"),
+            reader_options={"header": True},
+            splitter_method=None,
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 5
+    assert len(test_sparkdf.columns) == 2
+
+
+def test_get_batch_empty_splitter_tsv(
+    test_folder_connection_path_tsv, basic_spark_df_execution_engine
+):
+    # reader_method not configured because spark will configure own reader by default
+    # reader_options are needed to specify the fact that the first line of test file is the header
+    # reader_options are also needed to specify the separator (otherwise, comma will be used as the default separator)
+    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
+        PathBatchSpec(
+            path=os.path.join(test_folder_connection_path_tsv, "test.tsv"),
+            reader_options={"header": True, "sep": "\t"},
+            splitter_method=None,
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 5
+    assert len(test_sparkdf.columns) == 2
+
+
+def test_get_batch_empty_splitter_parquet(
+    test_folder_connection_path_parquet, basic_spark_df_execution_engine
+):
+    # Note: reader method and reader_options are not needed, because
+    # SparkDFExecutionEngine automatically determines the file type as well as the schema of the Parquet file.
+    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
+        PathBatchSpec(
+            path=os.path.join(test_folder_connection_path_parquet, "test.parquet"),
+            splitter_method=None,
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 5
+    assert len(test_sparkdf.columns) == 2
+
+
 def test_get_batch_with_split_on_whole_table_runtime(
     test_sparkdf, basic_spark_df_execution_engine
 ):
