+++ tests/execution_engine/split_and_sample/test_sqlalchemy_execution_engine_sampling.py
@@ -20,6 +20,11 @@ from great_expectations.execution_engine.sqlalchemy_dialect import GESqlDialect
 from great_expectations.self_check.util import build_sa_engine
 from great_expectations.util import import_library_module
 
+try:
+    sqlalchemy = pytest.importorskip("sqlalchemy")
+except ImportError:
+    sqlalchemy = None
+
 
 @pytest.mark.parametrize(
     "underscore_prefix",
@@ -72,7 +77,7 @@ def clean_query_for_comparison(query_string: str) -> str:
 @pytest.fixture
 def dialect_name_to_sql_statement():
     def _dialect_name_to_sql_statement(dialect_name: GESqlDialect) -> str:
-        DIALECT_NAME_TO_SQL_STATEMENT: dict = {
+        dialect_name_to_sql_statement: dict = {
             GESqlDialect.POSTGRESQL: "SELECT * FROM TEST_SCHEMA_NAME.TEST_TABLE WHERE TRUE LIMIT 10",
             GESqlDialect.MYSQL: "SELECT * FROM TEST_SCHEMA_NAME.TEST_TABLE WHERE TRUE = 1 LIMIT 10",
             GESqlDialect.ORACLE: "SELECT * FROM test_schema_name.test_table WHERE 1 = 1 AND ROWNUM <= 10",
@@ -87,7 +92,7 @@ def dialect_name_to_sql_statement():
             GESqlDialect.TRINO: "SELECT * FROM TEST_SCHEMA_NAME.TEST_TABLE WHERE TRUE LIMIT 10",
             GESqlDialect.HIVE: "SELECT * FROM `TEST_SCHEMA_NAME`.`TEST_TABLE` WHERE TRUE LIMIT 10",
         }
-        return DIALECT_NAME_TO_SQL_STATEMENT[dialect_name]
+        return dialect_name_to_sql_statement[dialect_name]
 
     return _dialect_name_to_sql_statement
 
@@ -149,15 +154,18 @@ def test_sample_using_limit_builds_correct_query_where_clause_none(
             #  and then use it here.
             dialect_name: GESqlDialect = self._dialect_name
             if dialect_name == GESqlDialect.ORACLE:
+                # noinspection PyUnresolvedReferences
                 return import_library_module(
                     module_name="sqlalchemy.dialects.oracle"
                 ).dialect()
             elif dialect_name == GESqlDialect.SNOWFLAKE:
+                # noinspection PyUnresolvedReferences
                 return import_library_module(
                     module_name="snowflake.sqlalchemy.snowdialect"
                 ).dialect()
             elif dialect_name == GESqlDialect.DREMIO:
                 # WARNING: Dremio Support is experimental, functionality is not fully under test
+                # noinspection PyUnresolvedReferences
                 return import_library_module(
                     module_name="sqlalchemy_dremio.pyodbc"
                 ).dialect()
@@ -168,11 +176,13 @@ def test_sample_using_limit_builds_correct_query_where_clause_none(
             #         module_name="sqlalchemy_redshift.dialect"
             #     ).RedshiftDialect
             elif dialect_name == GESqlDialect.BIGQUERY:
+                # noinspection PyUnresolvedReferences
                 return import_library_module(
                     module_name=self._BIGQUERY_MODULE_NAME
                 ).dialect()
             elif dialect_name == GESqlDialect.TERADATASQL:
                 # WARNING: Teradata Support is experimental, functionality is not fully under test
+                # noinspection PyUnresolvedReferences
                 return import_library_module(
                     module_name="teradatasqlalchemy.dialect"
                 ).dialect()
@@ -255,3 +265,90 @@ def test_sqlite_sample_using_limit(sa):
     for row_date in row_dates:
         assert row_date.month == 1
         assert row_date.year == 2018
+
+
+def test_sample_using_random(sqlite_view_engine, test_df):
+    my_execution_engine: SqlAlchemyExecutionEngine = SqlAlchemyExecutionEngine(
+        engine=sqlite_view_engine
+    )
+
+    p: float
+    batch_spec: SqlAlchemyDatasourceBatchSpec
+    batch_data: SqlAlchemyBatchData
+    num_rows: int
+    rows_0: List[tuple]
+    rows_1: List[tuple]
+
+    # First, make sure that degenerative case never passes.
+
+    test_df_0: pd.DataFrame = test_df.iloc[:1]
+    test_df_0.to_sql("test_table_0", con=my_execution_engine.engine)
+
+    p = 1.0
+    batch_spec = SqlAlchemyDatasourceBatchSpec(
+        table_name="test_table_0",
+        schema_name="main",
+        sampling_method="_sample_using_random",
+        sampling_kwargs={"p": p},
+    )
+
+    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
+    num_rows = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
+    ).scalar()
+    assert num_rows == round(p * test_df_0.shape[0])
+
+    rows_0: List[tuple] = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
+    ).fetchall()
+
+    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
+    num_rows = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
+    ).scalar()
+    assert num_rows == round(p * test_df_0.shape[0])
+
+    rows_1: List[tuple] = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
+    ).fetchall()
+
+    assert len(rows_0) == len(rows_1) == 1
+
+    assert rows_0 == rows_1
+
+    # Second, verify that realistic case always returns different random sample of rows.
+
+    test_df_1: pd.DataFrame = test_df
+    test_df_1.to_sql("test_table_1", con=my_execution_engine.engine)
+
+    p = 2.0e-1
+    batch_spec = SqlAlchemyDatasourceBatchSpec(
+        table_name="test_table_1",
+        schema_name="main",
+        sampling_method="_sample_using_random",
+        sampling_kwargs={"p": p},
+    )
+
+    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
+    num_rows = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
+    ).scalar()
+    assert num_rows == round(p * test_df_1.shape[0])
+
+    rows_0 = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
+    ).fetchall()
+
+    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
+    num_rows = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
+    ).scalar()
+    assert num_rows == round(p * test_df_1.shape[0])
+
+    rows_1 = batch_data.execution_engine.engine.execute(
+        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
+    ).fetchall()
+
+    assert len(rows_0) == len(rows_1)
+
+    assert not (rows_0 == rows_1)
