+++ tests/execution_engine/test_sqlalchemy_execution_engine.py
@@ -1,6 +1,5 @@
 import logging
 import os
-from typing import List
 
 import pandas as pd
 import pytest
@@ -12,9 +11,6 @@ from great_expectations.core.batch_spec import (
 )
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.execution_engine.execution_engine import MetricDomainTypes
-from great_expectations.execution_engine.sqlalchemy_batch_data import (
-    SqlAlchemyBatchData,
-)
 from great_expectations.execution_engine.sqlalchemy_execution_engine import (
     SqlAlchemyExecutionEngine,
 )
@@ -830,90 +826,3 @@ def test_sa_batch_unexpected_condition_temp_table(caplog, sa):
     )
 
     validate_tmp_tables()
-
-
-def test_sample_using_random(sqlite_view_engine, test_df):
-    my_execution_engine: SqlAlchemyExecutionEngine = SqlAlchemyExecutionEngine(
-        engine=sqlite_view_engine
-    )
-
-    p: float
-    batch_spec: SqlAlchemyDatasourceBatchSpec
-    batch_data: SqlAlchemyBatchData
-    num_rows: int
-    rows_0: List[tuple]
-    rows_1: List[tuple]
-
-    # First, make sure that degenerative case never passes.
-
-    test_df_0: pd.DataFrame = test_df.iloc[:1]
-    test_df_0.to_sql("test_table_0", con=my_execution_engine.engine)
-
-    p = 1.0
-    batch_spec = SqlAlchemyDatasourceBatchSpec(
-        table_name="test_table_0",
-        schema_name="main",
-        sampling_method="_sample_using_random",
-        sampling_kwargs={"p": p},
-    )
-
-    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
-    num_rows = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
-    ).scalar()
-    assert num_rows == round(p * test_df_0.shape[0])
-
-    rows_0: List[tuple] = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
-    ).fetchall()
-
-    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
-    num_rows = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
-    ).scalar()
-    assert num_rows == round(p * test_df_0.shape[0])
-
-    rows_1: List[tuple] = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
-    ).fetchall()
-
-    assert len(rows_0) == len(rows_1) == 1
-
-    assert rows_0 == rows_1
-
-    # Second, verify that realistic case always returns different random sample of rows.
-
-    test_df_1: pd.DataFrame = test_df
-    test_df_1.to_sql("test_table_1", con=my_execution_engine.engine)
-
-    p = 2.0e-1
-    batch_spec = SqlAlchemyDatasourceBatchSpec(
-        table_name="test_table_1",
-        schema_name="main",
-        sampling_method="_sample_using_random",
-        sampling_kwargs={"p": p},
-    )
-
-    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
-    num_rows = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
-    ).scalar()
-    assert num_rows == round(p * test_df_1.shape[0])
-
-    rows_0 = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
-    ).fetchall()
-
-    batch_data = my_execution_engine.get_batch_data(batch_spec=batch_spec)
-    num_rows = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.func.count()]).select_from(batch_data.selectable)
-    ).scalar()
-    assert num_rows == round(p * test_df_1.shape[0])
-
-    rows_1 = batch_data.execution_engine.engine.execute(
-        sqlalchemy.select([sqlalchemy.text("*")]).select_from(batch_data.selectable)
-    ).fetchall()
-
-    assert len(rows_0) == len(rows_1)
-
-    assert not (rows_0 == rows_1)
